+++ tests/helpers/templates/new/feature_src_layout.py
@@ -89,6 +89,7 @@ classifiers = [
   "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: Implementation :: CPython",
   "Programming Language :: Python :: Implementation :: PyPy",
 ]
@@ -103,9 +104,6 @@ Source = "https://github.com/unknown/{kwargs['project_name_normalized']}"
 [tool.hatch.version]
 path = "src/{kwargs['package_name']}/__about__.py"
 
-[tool.hatch.build.targets.sdist]
-[tool.hatch.build.targets.wheel]
-
 [tool.hatch.envs.default]
 dependencies = [
   "pytest",
@@ -116,7 +114,7 @@ cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=src/{k
 no-cov = "cov --no-cov"
 
 [[tool.hatch.envs.test.matrix]]
-python = ["37", "38", "39", "310"]
+python = ["37", "38", "39", "310", "311"]
 
 [tool.coverage.run]
 branch = true